gitignore issues fixed. Package moved up one level
