+++ src/chariot/api/database/DatabaseAPI.py
@@ -1,5 +1,5 @@
 from flask import Flask, request, app, jsonify, request
 
-from chariot.core.database.Database import Database
+from chariot.database.Database import Database
 
 DatabaseBaseUrl: str = '/chariot/api/v1.0/Database'
