+++ src/requirements.txt
@@ -1,0 +1,46 @@
+astroid==2.3.3
+attrs==19.3.0
+bandit==1.6.2
+Click==7.0
+entrypoints==0.3
+filelock==3.0.12
+flake8==3.7.9
+Flask==1.1.1
+gitdb2==2.0.6
+GitPython==3.0.5
+importlib-metadata==1.3.0
+isort==4.3.21
+itsdangerous==1.1.0
+Jinja2==2.10.3
+lazy-object-proxy==1.4.3
+MarkupSafe==1.1.1
+mccabe==0.6.1
+more-itertools==8.0.2
+mypy==0.761
+mypy-extensions==0.4.3
+packaging==20.0
+pbr==5.4.4
+pkg-resources==0.0.0
+pluggy==0.13.1
+py==1.8.1
+pycodestyle==2.5.0
+pyflakes==2.1.1
+pylint==2.4.4
+pyparsing==2.4.6
+pytest==5.3.2
+PyYAML==5.3
+six==1.13.0
+smmap2==2.0.5
+stevedore==1.31.0
+tinydb==3.15.2
+tinydb-serialization==1.0.4
+tinymongo==0.2.0
+toml==0.10.0
+tox==3.14.3
+typed-ast==1.4.0
+typing-extensions==3.7.4.1
+virtualenv==16.7.9
+wcwidth==0.1.8
+Werkzeug==0.16.0
+wrapt==1.11.2
+zipp==0.6.0
