+++ src/tox.ini
@@ -1,0 +1,16 @@
+[tox]
+envlist = py36, py37, py38
+
+[testenv]
+deps =
+    flake8
+    pylint
+    bandit
+    mypy
+    pytest
+commands = 
+    flake8 .
+    pylint .
+    bandit . -r
+    mypy .
+    pytest test -v
